added db func validation

+++ src/db/utils/crypto_fiat_trade_table.ts
@@ -1,23 +1,36 @@
 "use server";
 import { and, eq } from "drizzle-orm";
 import { db } from "../client";
-import { crypto_fiat_trade_table } from "../schema";
+import { crypto_fiat_trade_table, CryptoFiatTradeInsertT, CryptoFiatTradeT } from "../schema";
+import { createInsertSchema, createSelectSchema, createUpdateSchema } from "drizzle-zod";
+import { User } from "@supabase/supabase-js";
+import { getUser } from "@/lib/supabase/server";
+import { validateId } from "./validator";
 
-export type CryptoFiatTradeTableT = typeof crypto_fiat_trade_table.$inferSelect;
-export type CryptoFiatTradeTableInsertT = typeof crypto_fiat_trade_table.$inferInsert;
+export const createCryptoFiatTrade = async (trade: CryptoFiatTradeInsertT) => {
+	const user = await getUser();
+	if (!user) return null;
 
-export const createCryptoFiatTrade = async (trade: CryptoFiatTradeTableInsertT) => {
-	const res = await db.insert(crypto_fiat_trade_table).values(trade).returning().execute();
+	const validatedTradeData = validateInsert(user, trade);
+	if (!validatedTradeData) return null;
+
+	const res = await db.insert(crypto_fiat_trade_table).values(validatedTradeData).returning().execute();
 	if (res.length === 0) {
 		return null;
 	}
 	return res[0];
 }
 
-export const getCryptoFiatTrade = async (id: number, uid: string) => {
+export const getCryptoFiatTrade = async (id: number) => {
+	const user = await getUser();
+	if (!user) return null;
+
+	const validatedId = validateId(id);
+	if (!validatedId) return null;
+
 	const res = await db.select().from(crypto_fiat_trade_table).where(and(
 		eq(crypto_fiat_trade_table.id, id),
-		eq(crypto_fiat_trade_table.uid, uid)
+		eq(crypto_fiat_trade_table.uid, user.id)
 	)).execute();
 	if (res.length === 0) {
 		return null;
@@ -25,15 +38,27 @@ export const getCryptoFiatTrade = async (id: number, uid: string) => {
 	return res[0];
 }
 
-export const getAllCryptoFiatTrades = async (uid: string) => {
-	const res = await db.select().from(crypto_fiat_trade_table).where(eq(crypto_fiat_trade_table.uid, uid)).execute();
+export const getAllCryptoFiatTrades = async () => {
+	const user = await getUser();
+	if (!user) return null;
+
+	const res = await db.select().from(crypto_fiat_trade_table).where(eq(crypto_fiat_trade_table.uid, user.id)).execute();
 	return res;
 }
 
-export const updateCryptoFiatTrade = async (id: number, uid: string, trade: Partial<CryptoFiatTradeTableInsertT>) => {
-	const res = await db.update(crypto_fiat_trade_table).set(trade).where(and(
-		eq(crypto_fiat_trade_table.id, id),
-		eq(crypto_fiat_trade_table.uid, uid)
+export const updateCryptoFiatTrade = async (id: number, trade: Partial<CryptoFiatTradeInsertT>) => {
+	const user = await getUser();
+	if (!user) return null;
+
+	const validatedId = validateId(id);
+	if (!validatedId) return null;
+
+	const validatedTradeData = validateUpdate(user, trade);
+	if (!validatedTradeData) return null;
+
+	const res = await db.update(crypto_fiat_trade_table).set(validatedTradeData).where(and(
+		eq(crypto_fiat_trade_table.id, validatedId),
+		eq(crypto_fiat_trade_table.uid, user.id)
 	)).returning().execute();
 	if (res.length === 0) {
 		return null;
@@ -41,10 +66,16 @@ export const updateCryptoFiatTrade = async (id: number, uid: string, trade: Part
 	return res[0];
 }
 
-export const deleteCryptoFiatTrade = async (id: number, uid: string) => {
+export const deleteCryptoFiatTrade = async (id: number) => {
+	const user = await getUser();
+	if (!user) return null;
+
+	const validatedId = validateId(id);
+	if (!validatedId) return null;
+
 	const res = await db.delete(crypto_fiat_trade_table).where(and(
-		eq(crypto_fiat_trade_table.id, id),
-		eq(crypto_fiat_trade_table.uid, uid)
+		eq(crypto_fiat_trade_table.id, validatedId),
+		eq(crypto_fiat_trade_table.uid, user.id)
 	)).returning().execute();
 	if (res.length === 0) {
 		return null;
@@ -52,7 +83,44 @@ export const deleteCryptoFiatTrade = async (id: number, uid: string) => {
 	return res[0];
 }
 
-export const deleteAllCryptoFiatTrades = async (uid: string) => {
-	const res = await db.delete(crypto_fiat_trade_table).where(eq(crypto_fiat_trade_table.uid, uid)).returning().execute();
+export const deleteAllCryptoFiatTrades = async () => {
+	const user = await getUser();
+	if (!user) return null;
+
+	const res = await db.delete(crypto_fiat_trade_table).where(eq(crypto_fiat_trade_table.uid, user.id)).returning().execute();
 	return res;
+}
+
+const selectSchema = createSelectSchema(crypto_fiat_trade_table);
+const insertSchema = createInsertSchema(crypto_fiat_trade_table);
+const updateSchema = createUpdateSchema(crypto_fiat_trade_table);
+
+const validateSelect = (user: User, data: CryptoFiatTradeT) => {
+	const validatedData = selectSchema.safeParse(data);
+	if (!validatedData.success) {
+		console.error("Invalid trade data", validatedData.error.format());
+		return null;
+	}
+	validatedData.data.uid = user.id;
+	return validatedData.data;
+}
+
+const validateInsert = (user: User, data: CryptoFiatTradeInsertT) => {
+	const validatedData = insertSchema.safeParse(data);
+	if (!validatedData.success) {
+		console.error("Invalid trade data", validatedData.error.format());
+		return null;
+	}
+	validatedData.data.uid = user.id;
+	return validatedData.data;
+}
+
+const validateUpdate = (user: User, data: Partial<CryptoFiatTradeInsertT>) => {
+	const validatedData = updateSchema.safeParse(data);
+	if (!validatedData.success) {
+		console.error("Invalid trade data", validatedData.error.format());
+		return null;
+	}
+	validatedData.data.uid = user.id;
+	return validatedData.data;
 }
